+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 2.2.5 _(2021-09-21)_
 * New: Add DEVELOPER-DOCS.md and move hooks documentation into it
+* Change: Only support public post types, but also exclude 'attachment'
 * Change: Improve and tweak developer documentation and code examples
 * Change: Note compatibility through WP 5.8+
 * Unit tests: